iterar com while

+++ Iniciando_Python_Modulo1/FuncoesDeConstrucao.py
@@ -13,8 +13,8 @@ num1 = input('Digite um numero: ')
 num2 = input('Digite um numero: ')
 
 if num1.isdigit() and num2.isdigit():
-    num1 = int(num1)
-    num2 = int(num2)
+    num1 = float(num1)
+    num2 = float(num2)
 
     print(num1 + num2)
 else:
